Refactor Workload,Services,Istio Config List pages
