--- src/pages/ServiceList/ServiceListComponent.scss
@@ -1,12 +1,0 @@
-.ServiceList-Heading {
-  vertical-align: top;
-}
-
-.ServiceList-Title {
-  display: inline-block;
-}
-
-.health-icon {
-  top: 3px;
-  position: relative;
-}
