+++ src/pages/WorkloadList/WorkloadListPage.tsx
@@ -1,18 +1,201 @@
 import * as React from 'react';
 import * as FilterHelper from '../../components/FilterList/FilterHelper';
 import { RenderContent } from '../../components/Nav/Page';
-import WorkloadListContainer from './WorkloadListComponent';
 import * as WorkloadListFilters from './FiltersAndSorts';
+import * as FilterComponent from '../../components/FilterList/FilterComponent';
+import { WorkloadListItem, WorkloadNamespaceResponse } from '../../types/Workload';
+import { DurationInSeconds } from '../../types/Common';
+import Namespace from '../../types/Namespace';
+import { PromisesRegistry } from '../../utils/CancelablePromises';
+import { namespaceEquals } from '../../utils/Common';
+import { SortField } from '../../types/SortFilters';
+import { ActiveFiltersInfo } from '../../types/Filters';
+import { FilterSelected, StatefulFilters } from '../../components/Filters/StatefulFilters';
+import * as API from '../../services/Api';
+import VirtualList from '../../components/VirtualList/VirtualList';
+import TimeControlsContainer from '../../components/Time/TimeControls';
+import { KialiAppState } from '../../store/Store';
+import { activeNamespacesSelector, durationSelector } from '../../store/Selectors';
+import { connect } from 'react-redux';
+import DefaultSecondaryMasthead from '../../components/DefaultSecondaryMasthead/DefaultSecondaryMasthead';
 
-const WorkloadListPage: React.FunctionComponent = () => {
-  return (
-    <RenderContent>
-      <WorkloadListContainer
-        currentSortField={FilterHelper.currentSortField(WorkloadListFilters.sortFields)}
-        isSortAscending={FilterHelper.isCurrentSortAscending()}
-      />
-    </RenderContent>
-  );
+type WorkloadListPageState = FilterComponent.State<WorkloadListItem>;
+
+type ReduxProps = {
+  duration: DurationInSeconds;
+  activeNamespaces: Namespace[];
 };
+
+type WorkloadListPageProps = ReduxProps & FilterComponent.Props<WorkloadListItem>;
+
+class WorkloadListPageComponent extends FilterComponent.Component<
+  WorkloadListPageProps,
+  WorkloadListPageState,
+  WorkloadListItem
+> {
+  private promises = new PromisesRegistry();
+
+  constructor(props: WorkloadListPageProps) {
+    super(props);
+    const prevCurrentSortField = FilterHelper.currentSortField(WorkloadListFilters.sortFields);
+    const prevIsSortAscending = FilterHelper.isCurrentSortAscending();
+    this.state = {
+      listItems: [],
+      currentSortField: prevCurrentSortField,
+      isSortAscending: prevIsSortAscending
+    };
+  }
+
+  componentDidMount() {
+    this.updateListItems();
+  }
+
+  componentDidUpdate(prevProps: WorkloadListPageProps, _prevState: WorkloadListPageState, _snapshot: any) {
+    const prevCurrentSortField = FilterHelper.currentSortField(WorkloadListFilters.sortFields);
+    const prevIsSortAscending = FilterHelper.isCurrentSortAscending();
+    const [paramsSynced] = this.paramsAreSynced(prevProps);
+    if (!paramsSynced) {
+      this.setState({
+        currentSortField: prevCurrentSortField,
+        isSortAscending: prevIsSortAscending
+      });
+
+      this.updateListItems();
+    }
+  }
+
+  componentWillUnmount() {
+    this.promises.cancelAll();
+  }
+
+  paramsAreSynced = (prevProps: WorkloadListPageProps): [boolean, boolean] => {
+    const activeNamespacesCompare = namespaceEquals(prevProps.activeNamespaces, this.props.activeNamespaces);
+    const paramsSynced = prevProps.duration === this.props.duration && activeNamespacesCompare;
+    return [paramsSynced, activeNamespacesCompare];
+  };
+
+  sortItemList(workloads: WorkloadListItem[], sortField: SortField<WorkloadListItem>, isAscending: boolean) {
+    // Chain promises, as there may be an ongoing fetch/refresh and sort can be called after UI interaction
+    // This ensures that the list will display the new data with the right sorting
+    return this.promises.registerChained('sort', workloads, unsorted =>
+      WorkloadListFilters.sortWorkloadsItems(unsorted, sortField, isAscending)
+    );
+  }
+
+  updateListItems() {
+    this.promises.cancelAll();
+
+    const activeFilters: ActiveFiltersInfo = FilterSelected.getSelected();
+    const namespacesSelected = this.props.activeNamespaces.map(item => item.name);
+
+    if (namespacesSelected.length === 0) {
+      this.promises
+        .register('namespaces', API.getNamespaces())
+        .then(namespacesResponse => {
+          const namespaces: Namespace[] = namespacesResponse.data;
+          this.fetchWorkloads(
+            namespaces.map(namespace => namespace.name),
+            activeFilters
+          );
+        })
+        .catch(namespacesError => {
+          if (!namespacesError.isCanceled) {
+            this.handleAxiosError('Could not fetch namespace list', namespacesError);
+          }
+        });
+    } else {
+      this.fetchWorkloads(namespacesSelected, activeFilters);
+    }
+  }
+
+  getDeploymentItems = (data: WorkloadNamespaceResponse): WorkloadListItem[] => {
+    if (data.workloads) {
+      return data.workloads.map(deployment => ({
+        namespace: data.namespace.name,
+        name: deployment.name,
+        type: deployment.type,
+        appLabel: deployment.appLabel,
+        versionLabel: deployment.versionLabel,
+        istioSidecar: deployment.istioSidecar,
+        additionalDetailSample: deployment.additionalDetailSample,
+        healthPromise: API.getWorkloadHealth(
+          data.namespace.name,
+          deployment.name,
+          deployment.type,
+          this.props.duration,
+          deployment.istioSidecar
+        ),
+        labels: deployment.labels
+      }));
+    }
+    return [];
+  };
+
+  fetchWorkloads(namespaces: string[], filters: ActiveFiltersInfo) {
+    const workloadsConfigPromises = namespaces.map(namespace => API.getWorkloads(namespace));
+    this.promises
+      .registerAll('workloads', workloadsConfigPromises)
+      .then(responses => {
+        let workloadsItems: WorkloadListItem[] = [];
+        responses.forEach(response => {
+          workloadsItems = workloadsItems.concat(this.getDeploymentItems(response.data));
+        });
+        return WorkloadListFilters.filterBy(workloadsItems, filters);
+      })
+      .then(workloadsItems => {
+        this.promises.cancel('sort');
+        this.sortItemList(workloadsItems, this.state.currentSortField, this.state.isSortAscending)
+          .then(sorted => {
+            this.setState({
+              listItems: sorted
+            });
+          })
+          .catch(err => {
+            if (!err.isCanceled) {
+              console.debug(err);
+            }
+          });
+      })
+      .catch(err => {
+        if (!err.isCanceled) {
+          this.handleAxiosError('Could not fetch workloads list', err);
+        }
+      });
+  }
+
+  render() {
+    return (
+      <>
+        <div style={{ backgroundColor: '#fff' }}>
+          <DefaultSecondaryMasthead
+            rightToolbar={
+              <TimeControlsContainer
+                key={'DurationDropdown'}
+                id="workload-list-duration-dropdown"
+                handleRefresh={this.updateListItems}
+                disabled={false}
+              />
+            }
+          />
+        </div>
+        <RenderContent>
+          <VirtualList rows={this.state.listItems}>
+            <StatefulFilters
+              initialFilters={WorkloadListFilters.availableFilters}
+              onFilterChange={this.onFilterChange}
+            />
+          </VirtualList>
+        </RenderContent>
+      </>
+    );
+  }
+}
+
+const mapStateToProps = (state: KialiAppState) => ({
+  activeNamespaces: activeNamespacesSelector(state),
+  duration: durationSelector(state)
+});
+
+const WorkloadListPage = connect(mapStateToProps)(WorkloadListPageComponent);
 
 export default WorkloadListPage;
